// src/old/services/subsocial/squid/generated.ts
export type Maybe<T> = T | null;
export type InputMaybe<T> = Maybe<T>;
export type Exact<T extends { [key: string]: unknown }> = { [K in keyof T]: T[K] };
// ...
export type Scalars = {
  ID: { input: string | number; output: string; }
  String: { input: string; output: string; }
  Boolean: { input: boolean; output: boolean; }
  Int: { input: number; output: number; }
  Float: { input: number; output: number; }
  BigInt: { input: any; output: any; }
  DateTime: { input: any; output: any; }
  JSON: { input: any; output: any; }
};
// ...
export type AccountFollowers = {
  __typename?: 'AccountFollowers';
  followerAccount: Account;
  followingAccount: Account;
  id: Scalars['String']['output'];
};
// ...
export enum InReplyToKind {
  Post = 'Post'
}
// ...
export enum PostKind {
  Comment = 'Comment',
  RegularPost = 'RegularPost',
  SharedPost = 'SharedPost'
}
// ...
export type Query = {
  __typename?: 'Query';
  accountById?: Maybe<Account>;
  /** @deprecated Use accountById */
  accountByUniqueInput?: Maybe<Account>;
  accountFollowers: Array<AccountFollowers>;
  accountFollowersById?: Maybe<AccountFollowers>;
  /** @deprecated Use accountFollowersById */
  accountFollowersByUniqueInput?: Maybe<AccountFollowers>;
  accountFollowersConnection: AccountFollowersConnection;
  accounts: Array<Account>;
  accountsConnection: AccountsConnection;
  activeUsersTotalCount: ActiveUsersTotalCount;
  activeUsersTotalCountWithFilters: UserRetentionCountModel;
  activities: Array<Activity>;
  activitiesConnection: ActivitiesConnection;
  activityById?: Maybe<Activity>;
  /** @deprecated Use activityById */
  activityByUniqueInput?: Maybe<Activity>;
  commentFollowers: Array<CommentFollowers>;
  commentFollowersById?: Maybe<CommentFollowers>;
  /** @deprecated Use commentFollowersById */
  commentFollowersByUniqueInput?: Maybe<CommentFollowers>;
  commentFollowersConnection: CommentFollowersConnection;
  contentExtensionById?: Maybe<ContentExtension>;
  /** @deprecated Use contentExtensionById */
  contentExtensionByUniqueInput?: Maybe<ContentExtension>;
  contentExtensions: Array<ContentExtension>;
  contentExtensionsConnection: ContentExtensionsConnection;
  evmAccountById?: Maybe<EvmAccount>;
  /** @deprecated Use evmAccountById */
  evmAccountByUniqueInput?: Maybe<EvmAccount>;
  evmAccounts: Array<EvmAccount>;
  evmAccountsConnection: EvmAccountsConnection;
  evmSubstrateAccountLinkById?: Maybe<EvmSubstrateAccountLink>;
  /** @deprecated Use evmSubstrateAccountLinkById */
  evmSubstrateAccountLinkByUniqueInput?: Maybe<EvmSubstrateAccountLink>;
  evmSubstrateAccountLinks: Array<EvmSubstrateAccountLink>;
  evmSubstrateAccountLinksConnection: EvmSubstrateAccountLinksConnection;
  extensionPinnedResourceById?: Maybe<ExtensionPinnedResource>;
  /** @deprecated Use extensionPinnedResourceById */
  extensionPinnedResourceByUniqueInput?: Maybe<ExtensionPinnedResource>;
  extensionPinnedResources: Array<ExtensionPinnedResource>;
  extensionPinnedResourcesConnection: ExtensionPinnedResourcesConnection;
  inBatchNotifications: Array<InBatchNotifications>;
  inBatchNotificationsById?: Maybe<InBatchNotifications>;
  /** @deprecated Use inBatchNotificationsById */
  inBatchNotificationsByUniqueInput?: Maybe<InBatchNotifications>;
  inBatchNotificationsConnection: InBatchNotificationsConnection;
  ipfsFetchLogById?: Maybe<IpfsFetchLog>;
  /** @deprecated Use ipfsFetchLogById */
  ipfsFetchLogByUniqueInput?: Maybe<IpfsFetchLog>;
  ipfsFetchLogs: Array<IpfsFetchLog>;
  ipfsFetchLogsConnection: IpfsFetchLogsConnection;
  newsFeedById?: Maybe<NewsFeed>;
  /** @deprecated Use newsFeedById */
  newsFeedByUniqueInput?: Maybe<NewsFeed>;
  newsFeeds: Array<NewsFeed>;
  newsFeedsConnection: NewsFeedsConnection;
  notificationById?: Maybe<Notification>;
  /** @deprecated Use notificationById */
  notificationByUniqueInput?: Maybe<Notification>;
  notifications: Array<Notification>;
  notificationsConnection: NotificationsConnection;
  postById?: Maybe<Post>;
  /** @deprecated Use postById */
  postByUniqueInput?: Maybe<Post>;
  postFollowers: Array<PostFollowers>;
  postFollowersById?: Maybe<PostFollowers>;
  /** @deprecated Use postFollowersById */
  postFollowersByUniqueInput?: Maybe<PostFollowers>;
  postFollowersConnection: PostFollowersConnection;
  posts: Array<Post>;
  postsConnection: PostsConnection;
  reactionById?: Maybe<Reaction>;
  /** @deprecated Use reactionById */
  reactionByUniqueInput?: Maybe<Reaction>;
  reactions: Array<Reaction>;
  reactionsConnection: ReactionsConnection;
  searchQuery: ElasticSearchQueryResultEntity;
  spaceById?: Maybe<Space>;
  /** @deprecated Use spaceById */
  spaceByUniqueInput?: Maybe<Space>;
  spaceFollowers: Array<SpaceFollowers>;
  spaceFollowersById?: Maybe<SpaceFollowers>;
  /** @deprecated Use spaceFollowersById */
  spaceFollowersByUniqueInput?: Maybe<SpaceFollowers>;
  spaceFollowersConnection: SpaceFollowersConnection;
  spaces: Array<Space>;
  spacesConnection: SpacesConnection;
  squidStatus?: Maybe<SquidStatus>;
  userRetentionCount: UserRetentionCountFullModel;
};
// ...
export type Space = {
  __typename?: 'Space';
  /** The about text (bio) of a Space (IPFS content) */
  about?: Maybe<Scalars['String']['output']>;
  /** Is this a public space where anyone can post? */
  canEveryoneCreatePosts?: Maybe<Scalars['Boolean']['output']>;
  /** Are followers allowed to post in the Space? */
  canFollowerCreatePosts?: Maybe<Scalars['Boolean']['output']>;
  /** The CID of the content on IPFS */
  content?: Maybe<Scalars['String']['output']>;
  /** The block height when a Space was created. */
  createdAtBlock?: Maybe<Scalars['BigInt']['output']>;
  /** The DateTime when a Space was created. */
  createdAtTime?: Maybe<Scalars['DateTime']['output']>;
  /** A One-To-One relationship with the Account entity that created a Space. */
  createdByAccount: Account;
  /** The day when a Space was created. */
  createdOnDay?: Maybe<Scalars['DateTime']['output']>;
  /** The email address of a Space (IPFS content) */
  email?: Maybe<Scalars['String']['output']>;
  /** Space permissions rule */
  everyonePermissions?: Maybe<SpacePermissions>;
  /** The properties of a Space from its IPFS content which are not supported by the current squid's DB schema. */
  experimental?: Maybe<Scalars['JSON']['output']>;
  /** Space permissions rule */
  followerPermissions?: Maybe<SpacePermissions>;
  /** A Many-To-Many relationship between a Space and the Accounts that follow it through SpaceFollowers (foreign key - "followingSpace") */
  followers: Array<SpaceFollowers>;
  /** The total number of Accounts following a Space */
  followersCount: Scalars['Int']['output'];
  /** Space format (IPFS content) */
  format?: Maybe<Scalars['String']['output']>;
  /** The username of a Space (will be removed in further versions as it is deprecated. You should use the username field instead.) (IPFS content) */
  handle?: Maybe<Scalars['String']['output']>;
  /** Is the Space hidden? */
  hidden: Scalars['Boolean']['output'];
  /** The total number of hidden Posts in the current Space (post.length) */
  hiddenPostsCount: Scalars['Int']['output'];
  /** The ID of a Space, which will have the same value and Space ID on the blockchain. */
  id: Scalars['String']['output'];
  /** The URL of the Space's image (IPFS content) */
  image?: Maybe<Scalars['String']['output']>;
  /** A list of a Space's interests converted to a string with "comma" as a separator (IPFS content) */
  interestsOriginal?: Maybe<Scalars['String']['output']>;
  /** Is the Space's "About" section longer than its summary? */
  isShowMore?: Maybe<Scalars['Boolean']['output']>;
  /** A list of the Space's links converted to a string with "comma" as a separator (IPFS content) */
  linksOriginal?: Maybe<Scalars['String']['output']>;
  /** The name of a Space (IPFS content) */
  name?: Maybe<Scalars['String']['output']>;
  /** Space permissions rule */
  nonePermissions?: Maybe<SpacePermissions>;
  /** A One-To-One relationship with the Account entity that owns a Space. */
  ownedByAccount: Account;
  pinnedByExtensions: Array<ExtensionPinnedResource>;
  /** A One-To-Many relationship with the Posts created within the current Space (foreign key - "space") */
  posts: Array<Post>;
  /** The total number of all Posts (public and hidden) in the current Space (post.length) */
  postsCount: Scalars['Int']['output'];
  /** A source of profile data. */
  profileSource?: Maybe<Scalars['String']['output']>;
  /** A One-To-One relationship with the Account which uses the current Space as its profile. */
  profileSpace?: Maybe<Account>;
  /** The total number of public (non-hidden) Posts in the current Space (post.length) */
  publicPostsCount: Scalars['Int']['output'];
  /** Space permissions rule */
  spaceOwnerPermissions?: Maybe<SpacePermissions>;
  /** The summary of the content of a Space (IPFS content) */
  summary?: Maybe<Scalars['String']['output']>;
  /** A list of a Space's tags, converted to a string with "comma" as a separator (IPFS content) */
  tagsOriginal?: Maybe<Scalars['String']['output']>;
  /** The block height when a Space was updated. */
  updatedAtBlock?: Maybe<Scalars['BigInt']['output']>;
  /** The DateTime when a Space was updated. */
  updatedAtTime?: Maybe<Scalars['DateTime']['output']>;
  /** The username of a Space (IPFS content) */
  username?: Maybe<Scalars['String']['output']>;
};
// ...
export type GetPostsQueryVariables = Exact<{
  ids?: InputMaybe<Array<Scalars['String']['input']> | Scalars['String']['input']>;
}>;


export type GetPostsQuery = { __typename?: 'Query', posts: Array<{ __typename?: 'Post', content?: string | null, createdAtBlock?: any | null, createdAtTime?: any | null, title?: string | null, body?: string | null, summary?: string | null, isShowMore?: boolean | null, image?: string | null, link?: string | null, downvotesCount: number, hidden: boolean, id: string, isComment: boolean, kind?: PostKind | null, repliesCount: number, sharesCount: number, upvotesCount: number, updatedAtTime?: any | null, inReplyToKind?: InReplyToKind | null, followersCount: number, canonical?: string | null, tagsOriginal?: string | null, createdByAccount: { __typename?: 'Account', id: string }, inReplyToPost?: { __typename?: 'Post', id: string } | null, ownedByAccount: { __typename?: 'Account', id: string }, space?: { __typename?: 'Space', id: string } | null, rootPost?: { __typename?: 'Post', id: string, space?: { __typename?: 'Space', id: string } | null } | null, parentPost?: { __typename?: 'Post', id: string } | null, sharedPost?: { __typename?: 'Post', id: string } | null, extensions: Array<{ __typename?: 'ContentExtension', image?: string | null, amount?: any | null, chain?: string | null, collectionId?: string | null, decimals?: number | null, extensionSchemaId: ContentExtensionSchemaId, id: string, nftId?: string | null, token?: string | null, txHash?: string | null, message?: string | null, nonce?: string | null, url?: string | null, recipient?: { __typename?: 'Account', id: string } | null, fromEvm?: { __typename?: 'EvmAccount', id: string } | null, toEvm?: { __typename?: 'EvmAccount', id: string } | null, fromSubstrate?: { __typename?: 'Account', id: string } | null, toSubstrate?: { __typename?: 'Account', id: string } | null, pinnedResources: Array<{ __typename?: 'ExtensionPinnedResource', post?: { __typename?: 'Post', id: string } | null }> }> }> };

export type GetPostsFollowersCountQueryVariables = Exact<{
  ids?: InputMaybe<Array<Scalars['String']['input']> | Scalars['String']['input']>;
}>;


export type GetPostsFollowersCountQuery = { __typename?: 'Query', posts: Array<{ __typename?: 'Post', id: string, followersCount: number }> };

export type GetPostIdsBySpaceIdsQueryVariables = Exact<{
  ids?: InputMaybe<Array<Scalars['String']['input']> | Scalars['String']['input']>;
}>;


export type GetPostIdsBySpaceIdsQuery = { __typename?: 'Query', posts: Array<{ __typename?: 'Post', id: string, space?: { __typename?: 'Space', id: string } | null }> };